+++ app/codeInsee/SpecificProcessing.tsx
@@ -5,11 +5,11 @@ import { NormalizedInseeResult } from "./types";
 import { RowRecord } from "grist/GristData";
 import { WidgetColumnMap } from "grist/CustomSectionAPI";
 import { COLUMN_MAPPING_NAMES } from "./constants";
-import GenericChoiceBanner from "../../components/GenericChoiceBanner";
+import GenericChoiceBanner from "../../components/cleanData/GenericChoiceBanner";
 import { DEPT } from "../../lib/util/constants";
-import { DirtyRecord, NoResultRecord } from "../../lib/util/types";
+import { DirtyRecord, NoResultRecord } from "../../lib/cleanData/types";
 import RecordName from "../../components/RecordName";
-import GenericSpecificProcessing from "../../components/GenericSpecificProcessing";
+import GenericSpecificProcessing from "../../components/cleanData/GenericSpecificProcessing";
 
 export const SpecificProcessing: FC<{
   mappings: WidgetColumnMap | null;
