+++ app/codeInsee/page.tsx
@@ -19,7 +19,6 @@ import { Configuration } from "../../components/Configuration";
 import Image from "next/image";
 import globalSvg from "../../public/global-processing.svg";
 import specificSvg from "../../public/specific-processing.svg";
-import doneSvg from "../../public/done.svg";
 import { Instructions } from "./Instructions";
 import { SpecificProcessing } from "./SpecificProcessing";
 import {
@@ -28,9 +27,9 @@ import {
   NoResultRecord,
   UncleanedRecord,
   WidgetCleanDataSteps,
-} from "../../lib/util/types";
-import { cleanAndSortRecords } from "../../lib/util/utils";
-import GenericGlobalProcessing from "../../components/GenericGlobalProcessing";
+} from "../../lib/cleanData/types";
+import { cleanAndSortRecords } from "../../lib/cleanData/utils";
+import GenericGlobalProcessing from "../../components/cleanData/GenericGlobalProcessing";
 
 const InseeCode = () => {
   const [record, setRecord] = useState<RowRecord | null>();
