+++ app/geocode/SpecificProcessing.tsx
@@ -6,10 +6,10 @@ import { RowRecord } from "grist/GristData";
 import { WidgetColumnMap } from "grist/CustomSectionAPI";
 import { COLUMN_MAPPING_NAMES } from "./constants";
 import dynamic from "next/dynamic";
-import { DirtyRecord, NoResultRecord } from "../../lib/util/types";
-import GenericChoiceBanner from "../../components/GenericChoiceBanner";
+import { DirtyRecord, NoResultRecord } from "../../lib/cleanData/types";
+import GenericChoiceBanner from "../../components/cleanData/GenericChoiceBanner";
 import RecordName from "../../components/RecordName";
-import GenericSpecificProcessing from "../../components/GenericSpecificProcessing";
+import GenericSpecificProcessing from "../../components/cleanData/GenericSpecificProcessing";
 
 // react-leaflet is relies on browser APIs window. Dynamically load the component on the client side desabling ssr
 const MyAwesomeMap = dynamic(() => import("./Map"), { ssr: false });
